# packages/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: MOUSEKEEPER Control Plane API
  version: 0.1.0
  description: PostgreSQL에 먼저 영속되는 MOUSEKEEPER control-plane 공개 계약
security:
  - firebaseBearer: []
paths:
...
  /v1/notification-tokens:
    post:
      requestBody:
        required: true
        content:
          application/json:
            {
              schema:
                { $ref: "#/components/schemas/RegisterPushNotificationToken" },
            }
      responses:
        "201": { description: Mobile push token registered or refreshed }
  /v1/notification-tokens/{notificationTokenId}:
    parameters:
      - {
          in: path,
          name: notificationTokenId,
          required: true,
          schema: { type: string, format: uuid },
        }
    delete:
      responses:
        "200": { description: Mobile push token revoked }
  /v1/devices/{deviceId}/presence:
    parameters:
      - {
          in: path,
          name: deviceId,
          required: true,
          schema: { type: string, format: uuid },
        }
    get:
      responses:
        "200": { description: Current TTL-derived presence or OFFLINE }
  /v1/devices/{deviceId}/heartbeat:
    parameters:
      - {
          in: path,
          name: deviceId,
          required: true,
          schema: { type: string, format: uuid },
        }
    post:
      security: [{ deviceBearer: [] }]
      requestBody:
        required: true
        content:
          application/json:
            { schema: { $ref: "#/components/schemas/Heartbeat" } }
      responses:
        "201": { description: Presence TTL renewed for the exact device }
...
  /v1/rooms/{roomId}/commands:
    parameters:
      - {
          in: path,
          name: roomId,
          required: true,
          schema: { type: string, format: uuid },
        }
    get:
      responses:
        "200": { description: Room commands }
    post:
      parameters:
        - {
            in: header,
            name: Idempotency-Key,
            required: true,
            schema: { type: string, minLength: 8, maxLength: 128 },
          }
      requestBody:
        required: true
        content:
          application/json:
            { schema: { $ref: "#/components/schemas/CreateCommand" } }
      responses:
        "201": { description: Durable queued command }
  /v1/devices/{deviceId}/commands/pending:
    parameters:
      - {
          in: path,
          name: deviceId,
          required: true,
          schema: { type: string, format: uuid },
        }
    get:
      security: [{ deviceBearer: [] }]
      responses:
        "200":
          {
            description: Ordered commands not yet terminal for the exact desktop device,
          }
  /v1/devices/{deviceId}/commands/{commandId}/status:
    parameters:
      - {
          in: path,
          name: deviceId,
          required: true,
          schema: { type: string, format: uuid },
        }
      - {
          in: path,
          name: commandId,
          required: true,
          schema: { type: string, format: uuid },
        }
    patch:
      security: [{ deviceBearer: [] }]
      requestBody:
        required: true
        content:
          application/json:
            { schema: { $ref: "#/components/schemas/UpdateCommandStatus" } }
      responses:
        "200": { description: Validated command state transition }
...
  /v1/agent/proposals:
    post:
      security: [{ deviceBearer: [] }]
      parameters:
        - {
            in: header,
            name: Idempotency-Key,
            required: true,
            schema: { type: string, minLength: 8, maxLength: 128 },
          }
      requestBody:
        required: true
        content:
          application/json:
            { schema: { $ref: "#/components/schemas/CreateProposal" } }
      responses:
        "201": { description: Proposal and items persisted exactly once }
  /v1/proposals/{proposalId}:
    parameters:
      - {
          in: path,
          name: proposalId,
          required: true,
          schema: { type: string, format: uuid },
        }
    get:
      responses:
        "200":
          {
            description: Owned proposal with ordered items and README preview when present,
          }
  /v1/rooms/{roomId}/proposals/open:
    parameters:
      - {
          in: path,
          name: roomId,
          required: true,
          schema: { type: string, format: uuid },
        }
    get:
      responses:
        "200": { description: Open proposals for the owned room }
...
  /v1/rooms/{roomId}/file-browse-requests:
    parameters:
      - {
          in: path,
          name: roomId,
          required: true,
          schema: { type: string, format: uuid },
        }
    post:
      requestBody:
        required: true
        content:
          application/json:
            { schema: { $ref: "#/components/schemas/CreateFileBrowseRequest" } }
      responses:
        "201":
          { description: Browse request or immediate DEVICE_OFFLINE failure }
  /v1/file-browse-requests/{requestId}:
    parameters:
      - {
          in: path,
          name: requestId,
          required: true,
          schema: { type: string, format: uuid },
        }
    get:
      responses:
        "200": { description: Browse request state and completed page }
  /v1/devices/{deviceId}/file-browse-requests/pending:
    parameters:
      - {
          in: path,
          name: deviceId,
          required: true,
          schema: { type: string, format: uuid },
        }
    get:
      security: [{ deviceBearer: [] }]
      responses:
        "200":
          { description: Pending browse requests for the exact desktop device }
  /v1/agent/file-browse-requests/{requestId}/result:
    parameters:
      - {
          in: path,
          name: requestId,
          required: true,
          schema: { type: string, format: uuid },
        }
    post:
      security: [{ deviceBearer: [] }]
      requestBody:
        required: true
        content:
          application/json:
            { schema: { $ref: "#/components/schemas/CompleteFileBrowse" } }
      responses:
        "201": { description: Bounded relative-path browse page persisted }
  /v1/agent/file-browse-requests/{requestId}/failure:
    parameters:
      - {
          in: path,
          name: requestId,
          required: true,
          schema: { type: string, format: uuid },
        }
    post:
      security: [{ deviceBearer: [] }]
      requestBody:
        required: true
        content:
          application/json:
            { schema: { $ref: "#/components/schemas/FailFileBrowse" } }
      responses:
        "201": { description: Explicit browse failure persisted }
...
  /v1/rooms/{roomId}/chat:
    parameters:
      - {
          in: path,
          name: roomId,
          required: true,
          schema: { type: string, format: uuid },
        }
    get:
      responses:
        "200": { description: Up to 200 persisted room chat messages }
    post:
      requestBody:
        required: true
        content:
          application/json:
            { schema: { $ref: "#/components/schemas/CreateChatMessage" } }
      responses:
        "201":
          {
            description: User message persisted; AI remains UNCONFIGURED without a provider,
          }
...
components:
  securitySchemes:
    firebaseBearer:
      type: http
      scheme: bearer
      bearerFormat: Firebase ID token
    deviceBearer:
      type: http
      scheme: bearer
      bearerFormat: mk_device_ JWT
  schemas:
    RegisterPushNotificationToken:
      type: object
      additionalProperties: false
      required: [token, platform]
      properties:
        token: { type: string, minLength: 20, maxLength: 4096 }
        platform: { type: string, enum: [ANDROID, IOS] }
...
    Heartbeat:
      type: object
      additionalProperties: false
      required: [presence]
      properties:
        presence:
          {
            type: string,
            enum: [ONLINE_IDLE, ONLINE_SCANNING, ONLINE_EXECUTING, DEGRADED],
          }
...
    CreateCommand:
      oneOf:
        - type: object
          additionalProperties: false
          required: [intent, payload]
          properties:
            intent: { const: SCAN }
            payload:
              { type: object, additionalProperties: false, maxProperties: 0 }
        - type: object
          additionalProperties: false
          required: [intent, payload]
          properties:
            intent: { const: ANALYZE }
            payload:
              { type: object, additionalProperties: false, maxProperties: 0 }
        - type: object
          additionalProperties: false
          required: [intent, payload]
          properties:
            intent: { const: CREATE_RULE }
            payload:
              type: object
              additionalProperties: false
              required: [rule]
              properties:
                rule: { $ref: "#/components/schemas/CreateRule" }
        - type: object
          additionalProperties: false
          required: [intent, payload]
          properties:
            intent: { const: README }
            payload:
              type: object
              additionalProperties: false
              required: [purpose, audience, tone, sections]
              properties:
                purpose: { type: string, minLength: 1, maxLength: 500 }
                audience: { type: string, minLength: 1, maxLength: 200 }
                tone: { type: string, enum: [concise, friendly, technical] }
                sections:
                  type: array
                  maxItems: 20
                  items: { type: string, minLength: 1, maxLength: 120 }
    UpdateCommandStatus:
      type: object
      additionalProperties: false
      required: [status]
      properties:
        status: { type: string, enum: [DELIVERED, ANALYZING, FAILED] }
...
    ProposalItem:
      type: object
      additionalProperties: false
      required:
        [
          itemOrder,
          actionType,
          sourceRelativePath,
          destinationRelativePath,
          reasonCode,
          precondition,
          conflictState,
        ]
      properties:
        itemOrder: { type: integer, minimum: 0 }
        actionType:
          { type: string, enum: [MOVE, QUARANTINE, CREATE_DIR, README_WRITE] }
        sourceRelativePath: { type: [string, "null"], maxLength: 1024 }
        destinationRelativePath: { type: [string, "null"], maxLength: 1024 }
        reasonCode: { type: string, minLength: 1, maxLength: 100 }
        precondition: { type: object }
        conflictState:
          { type: string, enum: [NONE, NAME_CONFLICT, UNSUPPORTED] }
    CreateProposal:
      type: object
      additionalProperties: false
      required: [commandId, roomId, summary, expiresAt, items]
      properties:
        commandId: { type: string, format: uuid }
        roomId: { type: string, format: uuid }
        summary:
          type: object
          properties:
            itemCount: { type: integer, minimum: 0 }
            readmeDraft: { type: string, maxLength: 200000 }
            readmeDiff: { type: string, maxLength: 200000 }
        expiresAt: { type: [string, "null"], format: date-time }
        items:
          type: array
          minItems: 1
          maxItems: 200
          items: { $ref: "#/components/schemas/ProposalItem" }
...
    CreateFileBrowseRequest:
      type: object
      additionalProperties: false
      required: [relativeDirectory, cursor]
      properties:
        relativeDirectory: { type: string }
        cursor: { type: [string, "null"], maxLength: 512 }
    CompleteFileBrowse:
      type: object
      additionalProperties: false
      required: [entries, nextCursor, desktopGeneration]
      properties:
        entries:
          type: array
          maxItems: 200
          items:
            type: object
            additionalProperties: false
            required: [name, relativePath, type, sizeBytes, modifiedAt, fileId]
            properties:
              name: { type: string, minLength: 1, maxLength: 255 }
              relativePath: { type: string, minLength: 1, maxLength: 1024 }
              type: { type: string, enum: [FILE, DIRECTORY] }
              sizeBytes: { type: [integer, "null"], minimum: 0 }
              modifiedAt: { type: string, format: date-time }
              fileId: { type: string, minLength: 1, maxLength: 512 }
        nextCursor: { type: [string, "null"], maxLength: 512 }
        desktopGeneration: { type: string, minLength: 1, maxLength: 128 }
    FailFileBrowse:
      type: object
      additionalProperties: false
      required: [failureCode]
      properties:
        failureCode:
          {
            type: string,
            enum:
              [
                DEVICE_OFFLINE,
                TIMED_OUT,
                CURSOR_INVALIDATED,
                OUTSIDE_MANAGED_ROOT,
              ],
          }
...
    CreateChatMessage:
      type: object
      additionalProperties: false
      required: [content]
      properties:
        content: { type: string, minLength: 1, maxLength: 2000 }
...
    RuleDefinition:
      type: object
      additionalProperties: false
      required: [conditions, action]
      properties:
        match: { type: string, enum: [ALL, ANY], default: ALL }
        conditions:
          type: array
          minItems: 1
          maxItems: 10
          items:
            oneOf:
              - type: object
                additionalProperties: false
                required: [field, operator, value]
                properties:
                  field: { const: extension }
                  operator: { const: IN }
                  value:
                    {
                      type: array,
                      minItems: 1,
                      maxItems: 50,
                      items: { type: string, pattern: '^\\.[a-zA-Z0-9]+$' },
                    }
              - type: object
                additionalProperties: false
                required: [field, operator, value]
                properties:
                  field: { const: ageDays }
                  operator: { const: GTE }
                  value: { type: integer, minimum: 0, maximum: 36500 }
              - type: object
                additionalProperties: false
                required: [field, operator, value]
                properties:
                  field: { const: name }
                  operator:
                    { type: string, enum: [CONTAINS, STARTS_WITH, ENDS_WITH] }
                  value: { type: string, minLength: 1, maxLength: 120 }
        action:
          oneOf:
            - type: object
              additionalProperties: false
              required: [type, destinationTemplate]
              properties:
                type: { const: MOVE }
                destinationTemplate: { type: string, minLength: 1 }
            - type: object
              additionalProperties: false
              required: [type]
              properties:
                type: { const: QUARANTINE }
    CreateRule:
      type: object
      additionalProperties: false
      required: [name, definition, priority]
      properties:
        name: { type: string, minLength: 1, maxLength: 120 }
        definition: { $ref: "#/components/schemas/RuleDefinition" }
        priority: { type: integer, minimum: 0, maximum: 10000 }
        enabled: { type: boolean, default: true }
...